fix: circular import вынес NamedBytesIO в отдельный файл
